chore(@clayui/chart): update Chart component stories
